+++ common/predictive-text/unit_tests/headless/edit-distance/distance-modeler.js
@@ -7,7 +7,7 @@ function assertEdgeChars(edge, input, match) {
 }
 
 function edgeHasChars(edge, input, match) {
-  if(edge.optimalInput[edge.optimalInput.length - 1].sample.insert != input) {
+  if(edge.priorInput[edge.priorInput.length - 1].sample.insert != input) {
     return false;
   }
 
@@ -16,7 +16,7 @@ function edgeHasChars(edge, input, match) {
 
 function findEdgeWithChars(edgeArray, input, match) {
   let results = edgeArray.filter(function(value) {
-    return value.calculation.lastMatchEntry.key == match && value.optimalInput[value.optimalInput.length - 1].sample.insert == input;
+    return value.calculation.lastMatchEntry.key == match && value.priorInput[value.priorInput.length - 1].sample.insert == input;
   });
 
   assert.equal(results.length, 1);
@@ -47,9 +47,9 @@ describe('Correction Distance Modeler', function() {
 
         let childEdge = edges.filter(value => value.calculation.lastMatchEntry.key == child.char)[0];
         assert.isOk(childEdge);
-        assert.isEmpty(childEdge.optimalInput);
+        assert.isEmpty(childEdge.priorInput);
         assert.isEmpty(childEdge.calculation.inputSequence);
-        assert.equal(childEdge.currentCost, 1);
+        assert.isAbove(childEdge.currentCost, 0);
       }
 
       assert.equal(edges.length, expectedChildCount);
@@ -70,21 +70,23 @@ describe('Correction Distance Modeler', function() {
       let edges = rootNode.buildDeletionEdges(synthDistribution);
       assert.equal(edges.length, 2);
 
+      let highCost, lowCost;
       for(let edge of edges) {
-        assert.isNotEmpty(edge.optimalInput);
+        assert.isNotEmpty(edge.priorInput);
         assert.isEmpty(edge.calculation.matchSequence);
 
-        if(edge.optimalInput[0].p == 0.75) {
+        if(edge.priorInput[0].p == 0.75) {
           // 't'.
-          // 1.25:  assumes p = .75 ==> transform to input string cost of .25.
-          assert.equal(edge.currentCost, 1.25);
-        } else if(edge.optimalInput[0].p == 0.25) {
+          lowCost = edge.currentCost;  // higher prob = lower cost.
+        } else if(edge.priorInput[0].p == 0.25) {
           // 'h'
+          highCost = edge.currentCost;
         } else {
           assert.fail();
-          assert.equal(edge.currentCost, 1.75);
         }
       }
+
+      assert.isAbove(highCost, lowCost);
     });
 
     it('SearchNode.buildSubstitutionEdges() - from root', function() {
@@ -104,42 +106,53 @@ describe('Correction Distance Modeler', function() {
       assert.isAbove(edges.length, 0);
 
       let expectedChildCount = 0;
-      for(let child of rootTraversal.children()) {
-        for(let mass of synthDistribution) {
+      for(let mass of synthDistribution) { // probability may vary here
+        let highCost, lowCost;
+
+        // Within here, cost only varies based on edit-distance changes, allowing
+        // us to make an assertion on the relationship between costs for each edit type.
+        for(let child of rootTraversal.children()) {
           expectedChildCount++;
           
           let matchingEdge = findEdgeWithChars(edges, mass.sample.insert, child.char);
 
           // Substitution - matching char
           if(mass.sample.insert == matchingEdge.calculation.lastMatchEntry.key) {
-            assert.equal(matchingEdge.currentCost, 1 - mass.p);
+            if(lowCost) {
+              assert.equal(matchingEdge.currentCost, lowCost);
+            }
+            lowCost = matchingEdge.currentCost;
           } else {
             // not matching char.
-            assert.equal(matchingEdge.currentCost, 2 - mass.p);
+            if(highCost) {
+              assert.equal(matchingEdge.currentCost, highCost);
+            }
+            highCost = matchingEdge.currentCost;
           }
         }
+        assert.isAbove(highCost, lowCost);
       }
 
       assert.equal(edges.length, expectedChildCount);
 
       // One final bit, which is a bit of integration - we know the top two nodes that should result.
-      let queue = new models.PriorityQueue(correction.QUEUE_EDGE_COMPARATOR, edges);
+      let queue = new models.PriorityQueue(correction.QUEUE_NODE_COMPARATOR, edges);
 
       let firstEdge = queue.dequeue();
-      assert.equal(firstEdge.optimalInput[0].sample.insert, 't');
+      assert.equal(firstEdge.priorInput[0].sample.insert, 't');
       assert.equal(firstEdge.calculation.lastMatchEntry.key, 't');
-      assert.equal(firstEdge.currentCost, 0.25);
+      assert.isAbove(firstEdge.currentCost, 0);
 
       let secondEdge = queue.dequeue();
-      assert.equal(secondEdge.optimalInput[0].sample.insert, 'h');
+      assert.equal(secondEdge.priorInput[0].sample.insert, 'h');
       assert.equal(secondEdge.calculation.lastMatchEntry.key, 'h');
-      assert.equal(secondEdge.currentCost, 0.75);
+      assert.isAbove(secondEdge.currentCost, firstEdge.currentCost);
 
       // After this, a 't' input without a matching char.
       let nextEdge = queue.dequeue();
-      assert.equal(nextEdge.optimalInput[0].sample.insert, 't');
+      assert.equal(nextEdge.priorInput[0].sample.insert, 't');
       assert.notEqual(nextEdge.calculation.lastMatchEntry.key, 't');
-      assert.equal(nextEdge.currentCost, 1.25);
+      assert.isAbove(nextEdge.currentCost, secondEdge.currentCost);
     });
 
     it('Small integration test:  "teh" => "ten", "the"', function() {
@@ -166,13 +179,13 @@ describe('Correction Distance Modeler', function() {
       ];
 
       let layer1Edges = rootNode.buildSubstitutionEdges(synthDistribution1);
-      let layer1Queue = new models.PriorityQueue(correction.QUEUE_EDGE_COMPARATOR, layer1Edges);
+      let layer1Queue = new models.PriorityQueue(correction.QUEUE_NODE_COMPARATOR, layer1Edges);
 
       let tEdge = layer1Queue.dequeue();
       assertEdgeChars(tEdge, 't', 't');
 
-      let layer2Edges = new correction.SearchNode(rootTraversal, tEdge).buildSubstitutionEdges(synthDistribution2);
-      let layer2Queue = new models.PriorityQueue(correction.QUEUE_EDGE_COMPARATOR, layer2Edges);
+      let layer2Edges = tEdge.buildSubstitutionEdges(synthDistribution2);
+      let layer2Queue = new models.PriorityQueue(correction.QUEUE_NODE_COMPARATOR, layer2Edges);
 
       let eEdge = layer2Queue.dequeue();
       assertEdgeChars(eEdge, 'e', 'e');
@@ -186,40 +199,43 @@ describe('Correction Distance Modeler', function() {
       assert.isOk(ehEdge);
 
       // Final round:  we'll use three nodes and throw all of their results into the same priority queue.
-      let layer3eEdges  = new correction.SearchNode(rootTraversal, eEdge).buildSubstitutionEdges(synthDistribution3);
-      let layer3hEdges  = new correction.SearchNode(rootTraversal, hEdge).buildSubstitutionEdges(synthDistribution3);
-      let layer3ehEdges = new correction.SearchNode(rootTraversal, ehEdge).buildSubstitutionEdges(synthDistribution3);
-      let layer3Queue = new models.PriorityQueue(correction.QUEUE_EDGE_COMPARATOR, layer3eEdges.concat(layer3hEdges).concat(layer3ehEdges));
+      let layer3eEdges  = eEdge.buildSubstitutionEdges(synthDistribution3);
+      let layer3hEdges  = hEdge.buildSubstitutionEdges(synthDistribution3);
+      let layer3ehEdges = ehEdge.buildSubstitutionEdges(synthDistribution3);
+      let layer3Queue = new models.PriorityQueue(correction.QUEUE_NODE_COMPARATOR, layer3eEdges.concat(layer3hEdges).concat(layer3ehEdges));
 
       // Find the first result with an actual word directly represented.
+      let bestEdge;
       do {
-        edge = layer3Queue.dequeue();
-      } while(edge.calculation.lastMatchEntry.traversal.entries.length == 0);
+        bestEdge = layer3Queue.dequeue();
+      } while(bestEdge.calculation.lastMatchEntry.traversal.entries.length == 0);
 
-      assertEdgeChars(edge, 'n', 'n'); // 'ten' - perfect edit distance of 0, though less-likely input sequence.
-      assert(edge.currentCost, 1);
+      assertEdgeChars(bestEdge, 'n', 'n'); // 'ten' - perfect edit distance of 0, though less-likely input sequence.
+      // No cost assumptions here.
 
-      var edge;
+      var sibling1;
       do {
-        edge = layer3Queue.dequeue();
-      } while(edge.calculation.lastMatchEntry.traversal.entries.length == 0);
+        sibling1 = layer3Queue.dequeue();
+      } while(sibling1.calculation.lastMatchEntry.traversal.entries.length == 0);
 
       // Both have a raw edit distance of 1 while using the same input-sequence root. ('th')
-      let tenFlag = edgeHasChars(edge, 'h', 'n'); // subs out the 'h' entirely.  Could also occur with 'a', but is too unlikely.
-      let theFlag = edgeHasChars(edge, 'h', 'e'); // looks for transposed 'h' and 'e'.
+      let tenFlag = edgeHasChars(sibling1, 'h', 'n'); // subs out the 'h' entirely.  Could also occur with 'a', but is too unlikely.
+      let theFlag = edgeHasChars(sibling1, 'h', 'e'); // looks for transposed 'h' and 'e'.
 
       assert.isTrue(tenFlag || theFlag);
-      assert.equal(edge.currentCost, 1.5);
+      assert.isAbove(sibling1.currentCost, bestEdge.currentCost);
 
+      var sibling2;
       do {
-        edge = layer3Queue.dequeue();
-      } while(edge.calculation.lastMatchEntry.traversal.entries.length == 0);
+        sibling2 = layer3Queue.dequeue();
+      } while(sibling2.calculation.lastMatchEntry.traversal.entries.length == 0);
 
-      tenFlag = tenFlag || edgeHasChars(edge, 'h', 'n');
-      theFlag = theFlag || edgeHasChars(edge, 'h', 'e');
+      tenFlag = tenFlag || edgeHasChars(sibling2, 'h', 'n');
+      theFlag = theFlag || edgeHasChars(sibling2, 'h', 'e');
 
       assert.isTrue(tenFlag && theFlag);
-      assert.equal(edge.currentCost, 1.5);
+      assert.equal(sibling2.currentCost, sibling1.currentCost);
+      assert.isAbove(sibling2.currentCost, bestEdge.currentCost);
     });
   });
 
@@ -235,11 +251,9 @@ describe('Correction Distance Modeler', function() {
       assert.isFalse(firstSet.done);
       
       firstSet = firstSet.value; // Retrieves <actual value>
-      assert.equal(firstSet[1], 1);
-      assert.equal(firstSet[0].length, 1); // A single sequence ("ten") should be the best match.
-
-      let chars = firstSet[0][0].map(value => value.key);
-      assert.equal(chars.join(''), "ten");
+      // No checks on the first set's cost.
+      assert.equal(firstSet.length, 1); // A single sequence ("ten") should be the best match.
+      assert.equal(firstSet[0].matchString, "ten");
 
       let secondBatch = [
         'beh',  'te',  'tec',
@@ -252,12 +266,10 @@ describe('Correction Distance Modeler', function() {
       assert.isFalse(secondSet.done);
       
       secondSet = secondSet.value; // Retrieves <actual value>
-      assert.equal(secondSet[1], 1.5);
-      assert.equal(secondSet[0].length, secondBatch.length);
+      assert.isAbove(secondSet[0].totalCost, firstSet[0].totalCost); // assert it 'costs more'.
+      assert.equal(secondSet.length, secondBatch.length);
 
-      let entries = secondSet[0].map(function(sequence) {
-        return sequence.map(value => value.key).join('');
-      }).sort();
+      let entries = secondSet.map(result => result.matchString).sort();
       assert.deepEqual(entries, secondBatch);
 
       let thirdBatch = [
@@ -272,21 +284,31 @@ describe('Correction Distance Modeler', function() {
       assert.isFalse(thirdSet.done);
       
       thirdSet = thirdSet.value; // Retrieves <actual value>
-      assert.equal(thirdSet[1], 2);
-      assert.equal(thirdSet[0].length, thirdBatch.length);
+      assert.isAbove(thirdSet[0].totalCost, secondSet[0].totalCost); // assert it 'costs more'.
+      assert.equal(thirdSet.length, thirdBatch.length);
 
-      entries = thirdSet[0].map(function(sequence) {
-        return sequence.map(value => value.key).join('');
-      }).sort();
+      entries = thirdSet.map(result => result.matchString).sort();
       assert.deepEqual(entries, thirdBatch);
     }
+
+    it('Simple search without input', function() {
+      // The combinatorial effect here is a bit much to fully test.
+      let rootTraversal = testModel.traverseFromRoot();
+      assert.isNotEmpty(rootTraversal);
+
+      let searchSpace = new correction.SearchSpace(testModel);
+
+      let iter = searchSpace.getBestMatches();
+      let firstSet = iter.next();
+      assert.isFalse(firstSet.done);
+    });
 
     it('Simple search (paralleling "Small integration test")', function() {
       // The combinatorial effect here is a bit much to fully test.
       let rootTraversal = testModel.traverseFromRoot();
       assert.isNotEmpty(rootTraversal);
 
-      let searchSpace = new correction.SearchSpace(rootTraversal);
+      let searchSpace = new correction.SearchSpace(testModel);
 
       // VERY artificial distributions.
       let synthDistribution1 = [
@@ -307,7 +329,7 @@ describe('Correction Distance Modeler', function() {
       searchSpace.addInput(synthDistribution2);
       searchSpace.addInput(synthDistribution3);
 
-      let iter = searchSpace.getBestMatches();
+      let iter = searchSpace.getBestMatches(0); // disables the correction-search timeout.
       checkResults_teh(iter);
 
       // Debugging method:  a simple loop for printing out the generated sets, in succession.
@@ -337,7 +359,7 @@ describe('Correction Distance Modeler', function() {
       let rootTraversal = testModel.traverseFromRoot();
       assert.isNotEmpty(rootTraversal);
 
-      let searchSpace = new correction.SearchSpace(rootTraversal);
+      let searchSpace = new correction.SearchSpace(testModel);
 
       // VERY artificial distributions.
       let synthDistribution1 = [
@@ -358,13 +380,26 @@ describe('Correction Distance Modeler', function() {
       searchSpace.addInput(synthDistribution2);
       searchSpace.addInput(synthDistribution3);
 
-      let iter = searchSpace.getBestMatches();
+      let iter = searchSpace.getBestMatches(0); // disables the correction-search timeout.
       checkResults_teh(iter);
 
       // The key: do we get the same results the second time?
       // Reset the iterator first...
-      let iter2 = searchSpace.getBestMatches();
+      let iter2 = searchSpace.getBestMatches(0); // disables the correction-search timeout.
       checkResults_teh(iter2);
+    });
+
+    it('Empty search space, loaded model', function() {
+      // The combinatorial effect here is a bit much to fully test.
+      let rootTraversal = testModel.traverseFromRoot();
+      assert.isNotEmpty(rootTraversal);
+
+      let searchSpace = new correction.SearchSpace(testModel);
+      let iter = searchSpace.getBestMatches();
+
+      // While there's no input, insertion operations can produce suggestions.
+      let results = iter.next();
+      assert.isFalse(results.done);
     });
   });
 });