+++ windows/src/engine/kmcomapi/com/keyboardlanguages/keymankeyboardlanguage.pas
@@ -47,7 +47,9 @@ uses
   System.Variants,
   System.Win.ComObj,
 
+  BCP47Tag,
   glossary,
+  Keyman.System.Standards.LangTagsRegistry,
   keymanerrorcodes,
   utiltsf,
   utilxml;
@@ -56,12 +58,39 @@ uses
 
 constructor TKeymanKeyboardLanguage.Create(AContext: TKeymanContext;
   AOwner: IKeymanKeyboard; const ABCP47Code: string; ALangID: Integer; const AName: string);
+var
+  LangTag: TLangTag;
+  TempTag: string;
+  bcp47tag: TBCP47Tag;
 begin
   FContext := AContext;
   FOwner := AOwner;
   FBCP47Code := ABCP47Code;
   FLangID := ALangID;
-  FName := AName;
+
+  if AName <> '' then
+    // If the name is passed in, it comes from the package/keyboard metadata
+    FName := AName
+  else if TLangTagsMap.AllTags.TryGetValue(ABCP47Code, TempTag) and
+      TLangTagsMap.LangTags.TryGetValue(TempTag, LangTag) then
+    // No name is passed in, so we'll get it from langtags if possible
+    FName := LangTag.name
+  else
+  begin
+    // It's not a tag in langtags, so we'll lookup just the language subtag
+    bcp47tag := TBCP47Tag.Create(ABCP47Code);
+    try
+      if TLangTagsMap.AllTags.TryGetValue(bcp47tag.Language, TempTag) and
+          TLangTagsMap.LangTags.TryGetValue(TempTag, LangTag) then
+        FName := LangTag.name
+      else
+        // Still no luck, we'll just use the BCP47 tag
+        FName := ABCP47Code;
+    finally
+      bcp47tag.Free;
+    end;
+  end;
+
   inherited Create(AContext, IKeymanKeyboardLanguage);
 end;